+++ libs/ml/benchmark/embeddings/embeddings.cpp
@@ -21,6 +21,7 @@
 #include "ml/core/graph.hpp"
 #include "ml/layers/fully_connected.hpp"
 #include "ml/ops/activations/relu.hpp"
+#include "ml/ops/embeddings.hpp"
 #include "ml/ops/loss_functions/mean_square_error_loss.hpp"
 #include "ml/optimisation/adam_optimiser.hpp"
 #include "ml/optimisation/lazy_adam_optimiser.hpp"
