+++ libs/ml/examples/commodity_prediction/main.cpp
@@ -25,6 +25,8 @@
 #include "ml/exceptions/exceptions.hpp"
 #include "ml/layers/fully_connected.hpp"
 #include "ml/ops/activation.hpp"
+#include "ml/ops/activations/dropout.hpp"
+#include "ml/ops/activations/softmax.hpp"
 #include "ml/ops/loss_functions/mean_square_error_loss.hpp"
 #include "ml/optimisation/adam_optimiser.hpp"
