+++ libs/ml/src/optimisation/adagrad_optimiser.cpp
@@ -16,7 +16,12 @@
 //
 //------------------------------------------------------------------------------
 
+#include "ml/core/graph.hpp"
 #include "ml/optimisation/adagrad_optimiser.hpp"
+#include "ml/ops/trainable.hpp"
+#include "math/standard_functions/sqrt.hpp"
+#include "math/standard_functions/pow.hpp"
+#include "math/fundamental_operators.hpp"
 
 namespace fetch {
 namespace ml {
