+++ libs/ml/src/optimisation/momentum_optimiser.cpp
@@ -17,6 +17,8 @@
 //------------------------------------------------------------------------------
 
 #include "ml/optimisation/momentum_optimiser.hpp"
+#include "ml/core/graph.hpp"
+#include "ml/ops/trainable.hpp"
 
 namespace fetch {
 namespace ml {
