+++ libs/ml/src/optimisation/optimiser.cpp
@@ -17,9 +17,10 @@
 //------------------------------------------------------------------------------
 
 #include "ml/optimisation/optimiser.hpp"
-
-#include <chrono>
-#include <utility>
+#include "ml/core/graph.hpp"
+#include "math/standard_functions/pow.hpp"
+#include "ml/meta/ml_type_traits.hpp"
+#include "ml/ops/trainable.hpp"
 
 namespace fetch {
 namespace ml {
@@ -433,7 +434,7 @@ std::shared_ptr<Graph<TensorType>> Optimiser<TensorType>::GetGraph()
 ///////////////////////////////
 
 // TODO (ML-464)
-// template class Optimiser<math::Tensor<std::int8_t>>;
+//template class Optimiser<math::Tensor<std::int8_t>>;
 template class Optimiser<math::Tensor<std::int16_t>>;
 template class Optimiser<math::Tensor<std::int32_t>>;
 template class Optimiser<math::Tensor<std::int64_t>>;
